display right data on the dialog, now have to make it prettier

+++ app/src/main/java/com/weather/timeconverter/TimeConverter.kt
@@ -4,7 +4,7 @@ import java.text.SimpleDateFormat
 import java.util.*
 
 class TimeConverter {
-    fun getDateFromTime(time: Long): String {
+    fun getDateFromTime(time: Int): String {
         val dateFormat = SimpleDateFormat("EEEE d MMMM YYYY")
         val netDate = Date(time * 1000L)
         return dateFormat.format(netDate)
